requirements.txt
fastapi[standard]==0.115.12
uvicorn==0.34.2
beautifulsoup4==4.13.4
pyjwt==2.10.1
passlib[bcrypt]==1.7.4
requests==2.32.3
pandas==2.2.3
deep_translator==1.11.4
tenacity==9.1.2
pybreaker==1.3.0
python-json-logger==3.3.0
pytest==8.3.5